% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPSM.R
\name{bootPSM}
\alias{bootPSM}
\title{Run a complete parametric survival analysis}
\usage{
bootPSM(data, i, ...)
}
\arguments{
\item{data}{A data frame containing individual patient data for the relevant
time to event outcomes. This is passed to the \code{data} argument of the
\code{\link{runPSM}} function}

\item{i}{Index used to select a sample within \code{\link{boot}}.}

\item{...}{Additional parameters as used by \code{\link{runPSM}}}
}
\value{
The 'parameters_vector' object from the \code{\link{runPSM}} function.

'parameters_vector' is a vector which contains the coefficients for all of the flexsurv models specified.
The column names are in the format 'modeltype.distribution.parameter.TreatmentName', for example, comshp.weibull.shape.Int refers to the shape parameter
of the common shape weibull distribution for the intervention treatment and 'indshp.gengamma.mu.ref' refers to the mu parameter
of the independent shape generalised gamma distribution for the reference treatment. Columns with 'TE' at the end are the treatment effect coefficients
(applicable to the scale and shape parameters for independent shape models, applicable to the scale parameter only for the common shape
model and not applicable for the separate model).
}
\description{
This function is a wrapper for the \code{\link{runPSM}} function intended to be run with the \code{\link{boot}} package.
It enables running a complete parametric survival analysis for use when performing bootstrapping to explore uncertainty.
By re-using random seeds for each bootstrap sample it is possible to maintain correlations across multiple endpoints.
}
\details{
For more details and examples see the package vignettes:
\itemize{
\item \code{vignette("Fitting_models_in_R", package = "flexsurvPlus")}
\item \code{vignette("Fitting_models_in_R_bootstrap", package = "flexsurvPlus")}
\item \code{vignette("Survival_analysis_theory", package = "flexsurvPlus")}
}

This function is intended to be used in conjunction with the
\code{\link{boot}} function to return the statistic to be
bootstrapped. In this case by performing parametric survival modelling using
{\link{flexsurv}} and returning the parameters of the survival distributions.
This is used as the 'statistic' argument in the boot function.
}
