% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot strapping.R
\name{bootPSMtidy}
\alias{bootPSMtidy}
\title{Process bootstrap samples}
\usage{
bootPSMtidy(X)
}
\arguments{
\item{X}{The object created by running \code{\link{boot}} with \code{\link{runPSM}}}
}
\value{
A data frame similar to the 'parameters' object from the \code{\link{runPSM}} function.
}
\description{
This function does some minimal data manipulation to make the object returned by
bootPSM more useful and to match the \code{parameters} data frame that the \code{\link{runPSM}} function returns
}
\details{
This function is intended to be used in conjunction with the
\code{\link{boot}} function to return the statistic to be
bootstrapped. In this case by performing parametric survival modelling using
{\link{flexsurv}} and returning the parameters of the survival distributions.
This is used as the 'statistic' argument in the boot function.
}
