% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convSTEM.R
\name{convSTEM}
\alias{convSTEM}
\title{Convert survival parameters to SAS/STEM parametric forms}
\usage{
convSTEM(x = NULL, samples = NULL)
}
\arguments{
\item{x}{An object created by calling \code{\link{runPSM}}}

\item{samples}{An object created by calling \code{\link{boot}} with \code{\link{bootPSM}}

This function primarily exists for backward compatibility with older excel models where parametric extrapolation was
performed with SAS and alternative parametric forms were used for distributions. As such only a subset of models are supported.
One or both of \code{x} and \code{samples} must be specified and affect what is returned. For more details please see the
vignette \code{vignette("STEM_compatability", package = "flexsurvPlus")}

Possible distributions include

\itemize{
\item Exponential ('exp')
\item Weibull ('weibull')
\item Gompertz ('gompertz')
\item Log-normal ('lnorm')
\item Log-logistic ('llogis')
\item Generalized gamma ('gengamma')
\item Gamma ('gamma')
}}
}
\value{
a list containing 4 data frames
\itemize{
\item stem_param Converted parameter estimates
\item stem_cov Converted covariance matrix (if \code{samples} provided)
\item stem_modsum Converted model summary (if \code{x} provided)
\item stem_boot Converted bootstrap samples (if \code{samples} provided)
}
}
\description{
Convert survival parameters to SAS/STEM parametric forms
}
