% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit flexsurv models}
\usage{
fit_models(
  model.formula,
  distr = c("exp", "weibull", "gompertz", "lnorm", "llogis", "gengamma", "gamma",
    "genf"),
  data
)
}
\arguments{
\item{model.formula}{A survival model formula in the general form shown
below. Note that variable names must match the corresponding columns in the
data. This is passed to the \code{formula} argument of the
\code{\link{fit_models}} function
\itemize{
\item Surv(Time, Event==1) ~ ARM is a model with a single covariate for the
effect of treatment
\item Surv(Time, Event==1) ~ 1 is a model with no covariates typically
fitted to data from a single treatment group
}}

\item{data}{A data frame containing individual patient data for the relevant
time to event outcomes. This is passed to the \code{data} argument of the
\code{\link{fit_models}} function}

\item{dist}{A vector string of distributions, see dist argument in
\code{\link{flexsurvreg}}. This is passed to the \code{distr} argument of
the \code{\link{fit_models}} function}
}
\value{
A list containing flexsurv objects.
}
\description{
Fits single or multiple \code{\link{flexsurv}} models using \code{\link{flexsurvreg}}.
}
\details{
Possible distributions include:
\itemize{
\item Exponential ('exp')
\item Weibull ('weibull')
\item Gompertz ('gompertz')
\item Log-normal ('lnorm')
\item Log-logistic ('llogis')
\item Generalized gamma ('gengamma')
\item Gamma ('gamma')
\item Generalised F ('genf')
}
For more details and examples see the package vignettes:
\itemize{
\item \code{vignette("Fitting_models_in_R", package = "flexsurvPlus")}
\item \code{vignette("Fitting_models_in_R_bootstrap", package = "flexsurvPlus")}
\item \code{vignette("Survival_analysis_theory", package = "flexsurvPlus")}
}
}
\seealso{
\code{\link{flexsurvreg}}
}
