% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvPlus.R
\docType{package}
\name{flexsurvPlus}
\alias{flexsurvPlus}
\title{flexsurvPlus: A package for performing survival analyses.}
\description{
The flexsurvPlus package provides functions to help perform and summarize results
from basic survival analyses
}
