% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_curvefits.R
\name{get_curvefits}
\alias{get_curvefits}
\title{Get curve fits from \code{\link{fit_models}} objects}
\usage{
get_curvefits(models, time)
}
\arguments{
\item{models}{Object from \code{\link{fit_models}}}

\item{time}{list of times}
}
\value{
A data frame containing time and survival from flexsurv objects.
}
\description{
Manipulates \code{\link{fit_models}} objects to get the curve fits.
}
\seealso{
\code{\link{fit_models}} \code{\link{flexsurvreg}}
}
