% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_PSM.R
\name{get_model_summary}
\alias{get_model_summary}
\title{Get model summary from \code{\link{fit_models}} objects}
\usage{
get_model_summary(models)
}
\arguments{
\item{models}{Object from \code{\link{fit_models}}}
}
\value{
A data frame containing the \code{\link{AIC}}, \code{\link{BIC}}
from flexsurv objects and a status column of whether the model converged or produced an error or warning.
}
\description{
Manipulates \code{\link{fit_models}} objects to get a model summary of AIC and BIC and status of whether the model converged or produced an error or warning.
}
\seealso{
\code{\link{fit_models}} \code{\link{flexsurvreg}}
}
