% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_PSM.R
\name{get_params}
\alias{get_params}
\title{Get parameter estimates from \code{\link{fit_models}} objects}
\usage{
get_params(models)
}
\arguments{
\item{models}{Object from \code{\link{fit_models}}}
}
\value{
A data frame containing the \code{\link{AIC}}, \code{\link{BIC}}
from flexsurv objects and a status column of whether the model converged or produced an error or warning.
}
\description{
Manipulates \code{\link{fit_models}} objects to get the parameter estimates and AIC and BIC values.
}
\seealso{
\code{\link{fit_models}} \code{\link{flexsurvreg}}
}
