% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_independent_shape.R
\name{run_independent_shape}
\alias{run_independent_shape}
\title{Run a complete parametric survival analysis for an independent shape model}
\usage{
run_independent_shape(
  data,
  time_var,
  event_var,
  distr = c("weibull", "gompertz", "lnorm", "llogis", "gengamma", "gamma"),
  strata_var,
  int_name,
  ref_name
)
}
\arguments{
\item{data}{A data frame containing individual patient data for the relevant
time to event outcomes. This is passed to the \code{data} argument of the
\code{\link{fit_models}} function}

\item{time_var}{Name of time variable in 'data'. Variable must be numerical and >0.}

\item{event_var}{Name of event variable in 'data'. Variable must be
numerical and contain 1's to indicate an event and 0 to indicate a censor.}

\item{distr}{A vector string of distributions, see dist argument in
\code{\link{flexsurvreg}}. This is passed to the \code{distr} argument of
the \code{\link{fit_models}} function}

\item{strata_var}{Name of stratification variable in "data". This is usually
the treatment variable and must be categorical.}

\item{int_name}{Character to indicate the name of the treatment of interest,
must be a level of the "strata_var" column in "data", used for labeling
the parameters.
@param ref_name Character to indicate the name of the reference treatment,
must be a level of the "strata_var" column in "data", used for labeling
the parameters.}
}
\value{
A list containing 'models' (output from \code{\link{fit_models}}), 'model_summary' (output from\code{\link{get_params}}) and
'parameters', a data frame containing the coefficients of each flexsurv model.
\itemize{
\item 'models' is a list of flexsurv objects for each distribution specified
\item 'model_summary' is a tibble object containing the fitted model objects, the parameter
estimates (\code{\link{coef}}),  \code{\link{AIC}} and \code{\link{BIC}}
from flexsurv objects.
\item 'parameters' is a data frame with with one row which contains the coefficients for all of the flexsurv models specified.
The column names are in the format 'distribution.parameter.TreatmentName', for example, weibull.shape.Intervention refers to the shape parameter
of the weibull distribution for the intervention treatment and 'gengamma.mu.Reference' refers to the mu parameter
of the generalised gamma distribution for the reference treatment. Columns with 'TE' at the end are the treatment effect coefficients
(applicable to the scale and shape parameters for independent shape models).}
}
\description{
Fits \code{\link{flexsurv}} models using \code{\link{flexsurvreg}}
containing a covariate for treatment on shape and scale parameter.
}
\details{
Possible distributions include:
\itemize{
\item Exponential ('exp')
\item Weibull ('weibull')
\item Gompertz ('gompertz')
\item Log-normal ('lnorm')
\item Log-logistic ('llogis')
\item Generalized gamma ('gengamma')
\item Gamma ('gamma')
}

The model fit is in the form Surv(Time, Event==1) ~ ARM + shape(ARM). The
scale parameter is derived directly from the model for the reference
category, however for the intervention arm, this is calculated as reference
scale + treatment effect (scale). The shape parameter is derived directly
from the model for the reference category, however for the intervention
arm, this is calculated as reference shape + treatment effect (shape).
}
