% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_one_arm.R
\name{run_one_arm}
\alias{run_one_arm}
\title{Run a complete parametric survival analysis for one-arm}
\usage{
run_one_arm(
  data,
  time_var,
  event_var,
  distr = c("exp", "weibull", "gompertz", "lnorm", "llogis", "gengamma", "gamma",
    "genf"),
  int_name
)
}
\arguments{
\item{data}{A data frame containing individual patient data for the relevant
time to event outcomes. This is passed to the \code{data} argument of the
\code{\link{fit_models}} function}

\item{time_var}{Name of time variable in 'data'. Variable must be numerical and >0.}

\item{event_var}{Name of event variable in 'data'. Variable must be
numerical and contain 1's to indicate an event and 0 to indicate a censor.}

\item{distr}{A vector string of distributions, see dist argument in
\code{\link{flexsurvreg}}. This is passed to the \code{distr} argument of
the \code{\link{fit_models}} function. Default is all available distributions (see below).}

\item{int_name}{Character to indicate the name of the treatment of interest,
must be a level of the "strata_var" column in "data", used for labeling
the parameters.}
}
\value{
A list containing 'models' (output from \code{\link{fit_models}}), 'model_summary' (output from\code{\link{get_params}}) and
'parameters', a data frame containing the coefficients of each flexsurv model.
\itemize{
\item 'models' is a list of flexsurv objects for each distribution specified
\item 'model_summary' is a tibble object containing the fitted model objects, the parameter
estimates (\code{\link{coef}}),  \code{\link{AIC}} and \code{\link{BIC}}
from flexsurv objects.
\item 'parameters' is a data frame with with one row which contains the coefficients for all of the flexsurv models specified.
The column names are in the format 'distribution.parameter.TreatmentName', for example, weibull.shape.Intervention refers to the shape parameter
of the weibull distribution for the intervention treatment.}
}
\description{
Fits a single \code{\link{flexsurv}} models using \code{\link{flexsurvreg}}.
}
\details{
Possible distributions include:
\itemize{
\item Exponential ('exp')
\item Weibull ('weibull')
\item Gompertz ('gompertz')
\item Log-normal ('lnorm')
\item Log-logistic ('llogis')
\item Generalized gamma ('gengamma')
\item Gamma ('gamma')
\item Generalised F ('genf')
}

The model fit is in the form Surv(Time, Event==1) ~ 1 and is fit to the entire data (no strata). The parameters for each
treatment, are derived directly from the model (no additional manipulation
is required).
}
