% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PSM.R
\name{runPSM}
\alias{runPSM}
\title{Run a complete parametric survival analysis for one model with multiple distributions}
\usage{
runPSM(
  data,
  time_var,
  event_var,
  model.type = c("Separate", "Common shape", "Independent shape"),
  distr = c("exp", "weibull", "gompertz", "lnorm", "llogis", "gengamma", "gamma",
    "genf"),
  strata_var,
  int_name,
  ref_name
)
}
\arguments{
\item{data}{A data frame containing individual patient data for the relevant
time to event outcomes. This is passed to the \code{data} argument of the
\code{\link{fit_models}} function}

\item{time_var}{Name of time variable in 'data'. Variable must be numerical and >0.}

\item{event_var}{Name of event variable in 'data'. Variable must be
numerical and contain 1's to indicate an event and 0 to indicate a censor.}

\item{model.type}{Character vector indicating the types of model formula
provided. Permitted values are
\itemize{
\item 'Common shape' a model with a single covariate for the effect of
treatment on the scale parameter of the model
\item 'Independent shape' a model with a single covariate for treatment
that affects both the scale and shape parameters of the model
\item 'Separate' a model with no covariates typically fitted separately to
data from each treatment group in a study
\item 'One arm' a model with no covariates typically fitted to the entire data
set without a strata variable
}
Default is c("Separate", "Common shape", "Independent shape").}

\item{distr}{A vector string of distributions, see dist argument in
\code{\link{flexsurvreg}}. Default is all available distributions (see below).}

\item{strata_var}{Name of stratification variable in "data". This is usually
the treatment variable and must be categorical. Not required when model.type='One arm'.}

\item{int_name}{Character to indicate the name of the treatment of interest,
must be a level of the "strata_var" column in "data", used for labelling
the parameters.}

\item{ref_name}{Character to indicate the name of the reference treatment,
must be a level of the "strata_var" column in "data", used for labelling
the parameters. Not required when model.type='One arm'.}
}
\value{
A list containing 'models' (output from \code{\link{fit_models}}), 'model_summary' (output from \code{\link{get_params}}) and
'parameters', a data frame containing the coefficients of each flexsurv model.
\itemize{
\item 'models' is a list of flexsurv objects for each distribution specified
\item 'model_summary' is a tibble object containing the fitted model objects, the parameter
estimates (\code{\link{coef}}),  \code{\link{AIC}} and \code{\link{BIC}}
from flexsurv objects.
\item 'parameters' is a data frame with with one row per model.type called which contains the coefficients for all of the flexsurv models specified.
The column names are in the format 'distribution.parameter.TreatmentName', for example, weibull.shape.Intervention refers to the shape parameter
of the weibull distribution for the intervention treatment and 'gengamma.mu.Reference' refers to the mu parameter
of the generalised gamma distribution for the reference treatment. Columns with 'TE' at the end are the treatment effect coefficients
(applicable to the scale and shape parameters for independent shape models, applicable to the scale parameter only for the common shape
model and not applicable for the separate or one-arm model).
\item 'parameters_vector' is a vector which contains the coefficients for all of the flexsurv models specified.
The column names are in the format 'modeltype.distribution.parameter.TreatmentName', for example, comshp.weibull.shape.Int refers to the shape parameter
of the common shape weibull distribution for the intervention treatment and 'indshp.gengamma.mu.ref' refers to the mu parameter
of the independent shape generalised gamma distribution for the reference treatment. Columns with 'TE' at the end are the treatment effect coefficients
(applicable to the scale and shape parameters for independent shape models, applicable to the scale parameter only for the common shape
model and not applicable for the separate or one-arm model).
}
}
\description{
Run a complete parametric survival analysis for one model with multiple distributions
}
\details{
Possible distributions include:
\itemize{
\item Exponential ('exp')
\item Weibull ('weibull')
\item Gompertz ('gompertz')
\item Log-normal ('lnorm')
\item Log-logistic ('llogis')
\item Generalized gamma ('gengamma')
\item Gamma ('gamma')
\item Generalised F ('genf')
}
}
