% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adtte.R
\name{sim_adtte}
\alias{sim_adtte}
\title{A function to simulate Survival data}
\usage{
sim_adtte(
  rc_siminfo = FALSE,
  rc_origos = FALSE,
  id = 1,
  seed = 1234,
  rho = 0,
  pswitch = 0,
  proppd = 0,
  beta_1a = log(0.7),
  beta_1b = log(0.7),
  beta_2a = log(0.7),
  beta_2b = log(0.7),
  beta_pd = log(0.4),
  arm_n = 250,
  enroll_start = 0,
  enroll_end = 1,
  admin_censor = 2,
  os_gamma = 1.2,
  os_lambda = 0.3,
  ttp_gamma = 1.5,
  ttp_lambda = 2
)
}
\arguments{
\item{rc_siminfo}{Should simulation params be included in simulated dataframe (logical). Defaults to FALSE.}

\item{rc_origos}{Should OS without switching be included in simulated dataframe (logical). Defaults to FALSE.}

\item{id}{Identifer added to simulated dataframe. Defaults to 1.}

\item{seed}{Seed used for random number generator. Defaults to 1234.}

\item{rho}{correlation coefficient between TTP and OS. Defaults to 0.}

\item{pswitch}{proportion of patients who switch. Defaults to 0.}

\item{proppd}{proportion of patients with PFS before switch allowed. Defaults to 0.}

\item{beta_1a}{treatment effect (as log(Hazard Ratio)) for OS pre PFS. Defaults to log(0.7).}

\item{beta_1b}{treatment effect (as log(Hazard Ratio)) for OS post PFS. Defaults to log(0.7).}

\item{beta_2a}{treatment effect (as log(Hazard Ratio)) for OS pre PFS (switch). Defaults to log(0.7).}

\item{beta_2b}{treatment effect (as log(Hazard Ratio)) for OS post PFS (switch). Defaults to log(0.7).}

\item{beta_pd}{treatment effect on progression (as log(HR)). Defaults to log(0.4).}

\item{arm_n}{patients per arm. Defaults to 250.}

\item{enroll_start}{start of enrollment. Defaults to 0.}

\item{os_gamma}{weibull shape - for OS. Defaults to 1.2.}

\item{os_lambda}{weibull scale - for OS. Defaults to 0.3.}

\item{ttp_gamma}{weibull shape - for TTP. Defaults to 1.5.}

\item{ttp_lambda}{weibull scale - for TTP. Defaults to 2.}

\item{enroll.end}{end of enrollment. Defaults to 1.}

\item{admin.censor}{end of trial. Defaults to 2.}
}
\description{
This function simulates survival data with correlated time to progression and
overall survival times. Optionally crossover from treatment arms can be simulated.
}
\details{
The simulation times are derived from formulas in Austin 2012 adapted to enable correlations between endpoints.
Austin, P.C. (2012), Generating survival times to simulate Cox proportional hazards models with time‐varying covariates. Statist. Med., 31: 3946-3958. https://doi.org/10.1002/sim.5452
}
\examples{
require(survival)
require(dplyr)

ADTTE <- sim_adtte()
survfit(Surv(AVAL, event = CNSR ==  0) ~ ARM, data = filter(ADTTE, PARAMCD == "PFS")) \%>\%
  plot()

survfit(Surv(AVAL, event = CNSR ==  0) ~ ARM, data = filter(ADTTE, PARAMCD == "OS")) \%>\%
  plot()
}
\keyword{Simulation}
\keyword{Survival,}
