% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPSM.R
\name{summaryPSM}
\alias{summaryPSM}
\title{Extract information about fitted parametric survival models}
\usage{
summaryPSM(
  x,
  types = c("mean", "survival", "hazard", "cumhaz", "median", "rmst"),
  t = NULL,
  ci = FALSE,
  se = FALSE
)
}
\arguments{
\item{x}{An object created by calling \code{\link{runPSM}}}

\item{types}{A list of statistics to extract - see \code{\link{summary.flexsurvreg}} for details}

\item{t}{The time points to be used - see \code{\link{summary.flexsurvreg}} for details}

\item{ci}{Should a confidence interval be returned - see \code{\link{summary.flexsurvreg}} for details}

\item{se}{Should a standard error be returned - see \code{\link{summary.flexsurvreg}} for details}
}
\value{
A data frame containing the following values
\itemize{
\item Model - as specified in \code{\link{runPSM}} model.type
\item Dist - as specified in \code{\link{runPSM}} distr
\item Strata - Either Intervention or Reference
\item StrataName - As specified by int_name and ref_name respectively in \code{\link{runPSM}}
\item type - as defined by the types parameter see \code{\link{summary.flexsurvreg}} for details
\item variable - "est", "lcl", "ucl", "se" respectively
\item time - either NA or the time the statistic is evaluated at
\item value - estimated value
}
}
\description{
Extract information about fitted parametric survival models
}
\examples{

require(dplyr)
require(ggplot2)

PFS_data <- sim_adtte(seed = 2020, rho = 0.6) \%>\%
filter(PARAMCD=="PFS") \%>\%
transmute(USUBJID,
            ARMCD,
            PFS_days = AVAL,
            PFS_event = 1- CNSR
)

psm_pfs <- runPSM(
  data = PFS_data,
  time_var = "PFS_days",
  event_var = "PFS_event",
  strata_var = "ARMCD",
  int_name = "A",
  ref_name = "B")

summaryPSM(psm_pfs, types = c("mean","rmst"), t = c(100,2000)) \%>\%
   filter(Dist == "gengamma", Strata == "Intervention")

summaryPSM(psm_pfs, types = "survival", t = seq(0,2000,100)) \%>\%
  ggplot(aes(x=time, y = value, color = StrataName, linetype = Model)) +
  geom_line()+
  facet_grid(~Dist)

summaryPSM(psm_pfs, types = "hazard", t = seq(0,5000,100)) \%>\%
  ggplot(aes(x=time, y = value, color = StrataName, linetype = Model)) +
  geom_line()+
  facet_grid(~Dist) +
  coord_cartesian(ylim = c(0,0.02))

summaryPSM(psm_pfs, types = "cumhaz", t = seq(0,5000,100)) \%>\%
  ggplot(aes(x=time, y = value, color = StrataName, linetype = Model)) +
  geom_line()+
  facet_grid(~Dist) +
  coord_cartesian(ylim = c(0,100))

}
