% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_summary.R
\name{get_model_summary}
\alias{get_model_summary}
\title{Get model summary from \code{\link{fit_models}} objects}
\usage{
get_model_summary(models)
}
\arguments{
\item{models}{Object from \code{\link{fit_models}}}
}
\value{
A data frame containing the \code{\link{AIC}}, \code{\link{BIC}}
from flexsurv objects and a status column of whether the model converged or produced an error or warning.
}
\description{
Manipulates \code{\link{fit_models}} objects to get a model summary of AIC and BIC and status of whether the model converged or produced an error or warning.
}
\details{
For more details and examples see the package vignettes:
\itemize{
\item \code{vignette("Fitting_models_in_R", package = "flexsurvPlus")}
\item \code{vignette("Fitting_models_in_R_bootstrap", package = "flexsurvPlus")}
\item \code{vignette("Survival_analysis_theory", package = "flexsurvPlus")}
}
}
\seealso{
\code{\link{fit_models}} \code{\link{flexsurvreg}}
}
