% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_PSM.R
\name{get_params}
\alias{get_params}
\title{Get parameter estimates from \code{\link{fit_models}} objects}
\usage{
get_params(models)
}
\arguments{
\item{models}{Object from \code{\link{fit_models}}}
}
\value{
A tibble object containing the fitted model objects, the parameter
estimates (\code{\link{coef}}),  \code{\link{AIC}} and \code{\link{BIC}}
from flexsurv objects.
}
\description{
Manipulates \code{\link{fit_models}} objects to get the parameter estimates and AIC and BIC values.
}
\seealso{
\code{\link{fit_models}} \code{\link{flexsurvreg}}
}
